Correctly handle loading data from multiple chunks
This change enables the load method to work correctly when
data exists in multiple IDAT chunks. A small ByteReader helper
class is created that takes in a chunk iterator, and provides
a simple 'read' method. The reader code can then simply
read a scanline at a time and add to the canvas.